test(guard): preserve action class contracts
